PMFS: Don't override root inode perms on mount
For some reason we were manually overriding the root inode permissions
to be drwxrwxrwt on each mount.  This commit takes that code out and
leaves the permissions as they were on unmount.

I also changed an error message that was erroneously saying we were
doing something to recover when the root inode wasn't marked as being a
directory.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Acked-by: Rob Gittins <rob.gittins@intel.com>