PMFS: Fix race with inode allocation/eviction
This change fixes a race that existed between inode allocation and eviction.
Prior to this change we could get into a situation where we were trying to free
an old inode's data blocks based on information that had changed since the
inode was reallocated.

With this change we calculate and save everything we need to free the
data blocks before we release the inode.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Reviewed-by: Sanjay Kumar <sanjay.k.kumar@intel.com>