PMFS: Optimize freeing of PMFS data blocks
This change optimizes iterative calls to pmfs_free_block so that
subsequent calls begin the search for their block where the last search
terminated.  Prior to this change the search was done from the head of
the list, which resulted in an O(n^2) algorithm in the worst case.  This
optimization relies on the fact that block numbers allocated to a given
file are normally in strictly ascending order.

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Reviewed-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
Reported-by: Boris Zuckerman <borisz@hp.com>