PMFS: Fix a couple of bugs in pmfs_inode_count_iblocks()
pmfs_inode_count_iblocks_recursive() was taking its 'block' number
as an unsigned long, which would be truncated on a 32-bit system.
Change it to take a __le64 instead, and adjust the callers.  This fixes
an endianness bug as the caller of pmfs_inode_count_iblocks() was not
swapping its argument.

I could have made a smaller change here by leaving 'root' as u64, and
adding byteswapping to the caller, but this way feels more similar to
the way the rest of the file handles block types.

Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>