PMFS: Fix endian bugs in journal.c
The code in pmfs_clean_journal() assumed that the CPU was little-endian
(ie that reading a 64-bit quantity and then masking it was 2^31-1 would
return the data in the lower 32-bits of storage).  Make that assumption
true by swapping the full 64-bit quantity read, and then shifting/masking
to extract the various components of it.

Fixing the code which writes this location is even more interesting
because we want to write the 64-bit quantity atomically, using
set_64bit().  Unfortunately, set_64bit() does not take a __le64 but a
u64, so first we calculate the __le64 value we want, then we throw away
all the type checking by using __force.  Not my favourite way to fix
the problem, but I didn't want to add a set_64bit_le or similar.

Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>