PMFS: Enhancements to page zeroing path
With the current tests for zeroing a page, there are cases where we zero
pages twice - if the start page and end page happen to be the same, and
both are 'edge' conditions.

This change fixes that, and further, instead of zeroing the whole page,
zeroes only the part of the page which is not going to be written to.

Signed-off-by: Vishal Verma <vishal.l.verma@linux.intel.com>
Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>