pmfs: Remove ioremap-related sparse warnings
ioremap returns an __iomem pointer which should not normally be directly
dereferenced.  The pmfs filesystem wants a pointer that can be directly
dereferenced; the use of ioremap to obtain that pointer is a temporary
implementation artefact which will eventually go away.  Disable the sparse
warnings by inserting the __force casts at the minimum number of places.

Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>